+++ controllers/kafkatopic_controller.go
@@ -53,7 +53,6 @@ func (h KafkaTopicHandler) createOrUpdate(ctx context.Context, avn *aiven.Client
 
 	exists, err := h.exists(ctx, avn, topic)
 	if err != nil {
-		meta.SetStatusCondition(&topic.Status.Conditions, getErrorCondition("CheckExists", err))
 		return err
 	}
 
@@ -67,7 +66,6 @@ func (h KafkaTopicHandler) createOrUpdate(ctx context.Context, avn *aiven.Client
 			Config:      convertKafkaTopicConfig(topic),
 		})
 		if err != nil && !aiven.IsAlreadyExists(err) {
-			meta.SetStatusCondition(&topic.Status.Conditions, getErrorCondition("Create", err))
 			return err
 		}
 
@@ -81,7 +79,6 @@ func (h KafkaTopicHandler) createOrUpdate(ctx context.Context, avn *aiven.Client
 				Config:      convertKafkaTopicConfig(topic),
 			})
 		if err != nil {
-			meta.SetStatusCondition(&topic.Status.Conditions, getErrorCondition("Update", err))
 			return fmt.Errorf("cannot update Kafka Topic: %w", err)
 		}
 
@@ -115,7 +112,6 @@ func (h KafkaTopicHandler) delete(ctx context.Context, avn *aiven.Client, obj cl
 	// Delete project on Aiven side
 	err = avn.KafkaTopics.Delete(ctx, topic.Spec.Project, topic.Spec.ServiceName, topic.GetTopicName())
 	if err != nil && !aiven.IsNotFound(err) {
-		meta.SetStatusCondition(&topic.Status.Conditions, getErrorCondition("Delete", err))
 		return false, err
 	}
 
@@ -172,11 +168,7 @@ func (h KafkaTopicHandler) checkPreconditions(ctx context.Context, avn *aiven.Cl
 	meta.SetStatusCondition(&topic.Status.Conditions,
 		getInitializedCondition("Preconditions", "Checking preconditions"))
 
-	running, err := checkServiceIsRunning(ctx, avn, topic.Spec.Project, topic.Spec.ServiceName)
-	if err != nil {
-		meta.SetStatusCondition(&topic.Status.Conditions, getErrorCondition("Preconditions", err))
-	}
-	return running, err
+	return checkServiceIsRunning(ctx, avn, topic.Spec.Project, topic.Spec.ServiceName)
 }
 
 func (h KafkaTopicHandler) getState(ctx context.Context, avn *aiven.Client, topic *v1alpha1.KafkaTopic) (string, error) {